+++ fixtures/coverall/tests/bindings/test_coverall.kts
@@ -138,6 +138,9 @@ Coveralls("test_complex_errors").use { coveralls ->
     } catch(e: ComplexException.OsException) {
         assert(e.code == 10.toShort())
         assert(e.extendedCode == 20.toShort())
+        assert(e.toString() == "uniffi.coverall.ComplexException\$OsException: code=10, extendedCode=20") {
+            "Unexpected ComplexError.OsError.toString() value: ${e.toString()}"
+        }
     }
 
     try {
@@ -145,6 +148,9 @@ Coveralls("test_complex_errors").use { coveralls ->
         throw RuntimeException("Expected method to throw exception")
     } catch(e: ComplexException.PermissionDenied) {
         assert(e.reason == "Forbidden")
+        assert(e.toString() == "uniffi.coverall.ComplexException\$PermissionDenied: reason=Forbidden") {
+            "Unexpected ComplexError.PermissionDenied.toString() value: ${e.toString()}"
+        }
     }
 
     try {